app/requirements.txt
appdirs==1.4.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Login==0.5.0
idna==2.10
importlib-metadata==3.3.0
importlib-resources==4.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
pipenv==2020.11.15
pycparser==2.20
pyOpenSSL==20.0.1
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
Werkzeug==1.0.1
zipp==3.4.0
